cmd: usb: Update command
The function usb_control_msg returns negative number for error, but here
it was checking only for -1, so it was missing different errors. So if we
run the command 'usb test 0 0 J' then it would fail but it would return
-ENODEV meaning that it would say that it succeed.
Also update the print on the success branch, to make it clear that first
it needs to stop before starting again.