sparx5: Update board_init
On PCB135 on latest version (at least version 6) it has Indy PHYs. And
to get these PHY functional it is required to release the reset to the
PHYs and then also to release the coma mode. The reset is on GPIO 19
while the coma mode is on GPIO 16. The GPIO 19 was already released but
the GPIO 16 wasn't released. Therefore fix this setting GPIO 16 as
output.
We do similar for the linux kernel regarding of the previous PCB135
versions where were different PHYs. It seems that it doesn't have any
side effects.