sparx5_switch: When using ports 12-15 they need to set in SGMII mode
When configuring the ports 12-15 for using the QSGMII interface it is
required to set the register PORT_CONF_DEV10G_MODES.

With this change it is possible to enable the front ports on pcb135 for
network usage. The change needs to be done in the device tree. Here is
an example where port0 is enabled:

&mdio0 {
	status = "okay";
	phy0: ethernet-phy@0 {
		reg = <0>;
	};
};

&switch {
	status = "okay";
	ethernet-ports {
		port0: port@0 {
			reg = <0>;
			phy-handle = <&phy0>;
			phys = <IF_QSGMII FA_SERDES_TYPE_10G 13>; // type serdes serdes#
		};
	};
};