sparxt_serdes: Fix support for IF_SGMII_CISCO
When the interface IF_SGMII_CISCO is used then the serdes was not doing
anything because it didn't understand that interface. Fix this by
setting the IF_SGMII_CISCO to behave as the IF_SGMII interface.
The IF_SGMII_CISCO is used for the CuSFPs. Currently we don't talk with
the PHYs of the CuSFP and we just enabled aneg between the MAC and PHY.
This limits the functionality to run only at 1G.