sparx5: lan969x: Enable all the ports to run at lower speed or equal to 2.5G
As we currently don't support higher speeds in u-boot, we enable all the
ports to run at speeds lower than 2.5G. With this we can enable the
IF_SGMII_CISCO on the SFP ports. The problem was that those chip ports
were not configured to run at lower speeds so the PCS will not be able
to auto negotiate with the PHY. By putting all the ports at lower speeds
then the MAC can autonegotiate with the PHY and then traffic is passing
through.

For example to enable the first SFP port on ev23x71a:

&switch {
	status = "okay";
	ethernet-ports {
		port24: port@24 {
			reg = <24>;
			phys = <IF_SGMII_CISCO FA_SERDES_TYPE_10G 6>;
		};
	};
};