sparx5: Get network working also on 25G Serdes
By default the 25G ports that are connected to the 25G serdes were not
enabled to run at lower speeds. Now allow also these ports to run at
lower speeds to be able to use them also in u-boot.
With this change it is possible to update the device tree to set these
ports to be used as network management ports. By default still the
port64 is used for management. To change this the following change needs
to be done in the device tree:

&switch {
	status = "okay";
	ethernet-ports {
		port56: port@56 {
			reg = <56>;
			phys = <IF_SGMII_CISCO FA_SERDES_TYPE_25G 25>; // type serdes serdes#
		};
	};
};

This will enable the CuSFP ports to be used.