configs: mchp_lan969x_sram_defconfig: Shrink the environment size
Apparently when we try to save the env in u-boot we allocate memory in
SRAM/DRAM to be able to read the entire env. But because we are limited
by space when running from SRAM, we don't have enough memory, so the save
command will fail.
To fix this for now decrease the env size by half for both the env and
for redundant env.