net: sparx5: Allow the RGMII ports to run at lower speeds
Allow the RGMII ports to run at lower speeds. Basically the PHY will
autonegotiate with the partner and if they autonegotiate at a speed of
100Mbit then the traffic was not passing.
The issue was that the RGMII ports were configured to run at 1G
regardless of why the PHY autonegotiate.
One important change here is that we configure again the port after the
PHY has finished the autonegotiation.