board: lan969x: ev09p11a/ev22n59a: Enable DDR_PLL
On the board ev22n59a there were problems with the clock towards the
Watson PHYs. To fix this the HW have done some ECOs and with these ECOs
the SW needs to be updated. The SW needs to configure the DDR_PLL and
set the GPIO 10 as alternate mode 3 and GPIO 9 as input enabled to
disable the MDIO0. If these changes are not made then the NPI port
doesn't work.